refactor: make report as submodule
